replace cancancan with pundit

+++ app/graphql/mutations/base_remove_mutation.rb
@@ -5,13 +5,15 @@ module Mutations
     class_attribute :record_field_name,
                     :entity_klass_name,
                     :entity_type,
-                    :model_klass
+                    :model_klass,
+                    :policy
 
     def self.set_entity
       self.entity_klass_name ||= to_s.demodulize.gsub('Remove', '')
       self.record_field_name ||= entity_klass_name.to_s.underscore.to_sym
       self.entity_type ||= Types.const_get("#{entity_klass_name}Type")
       self.model_klass ||= entity_klass_name.constantize
+      self.policy ||= "#{entity_klass_name}Policy".constantize
 
       description "Remove #{entity_klass_name} from database"
 
@@ -27,17 +29,15 @@ module Mutations
       @current_user ||= context[:current_user]
     end
 
-    def current_ability
-      @current_ability ||= Ability.new(current_user)
-    end
-
     def resolve(id:)
       record = self.class.model_klass.find(id)
-      current_ability.authorize! :destroy, record
 
-      record.destroy!
-
-      { self.class.record_field_name => record }
+      if policy.new(current_user, record).destroy?
+        record.destroy!
+        { self.class.record_field_name => record }
+      else
+        GraphQL::ExecutionError.new('You are not allowed to perform this action')
+      end
     rescue ActiveRecord::RecordNotDestroyed => e
       GraphQL::ExecutionError.new(e.record.errors.full_messages.first)
     end
